# Optimize `Object::cast_to` for final classes by using `typeid` instead of `dynamic_cast`.