@lawnjelly 2025-03-06
> Also, it is unclear to me whether dynamic_cast is used in any performance critical functions.

`Object::cast_to` is used all over the shop, including traversing scene trees, so it should be fast.